Fix and use ovn probe
Currently there is a bug in the ovn probe. When the go-ovn client
is created it scans all the resources and makes the callbacks.
These callbacks emit a message in the channel for the processing
goroutine to pick up and handle.

Since the processing goroutine has not started yet, if the channel
gets full, the callbacks block and the go-ovn client is never created.

This needs to be fixed by first creating the processing goroutine.

As a workaround for now, increase the channel size.

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>