ovn-probe: Create worker goroutine before ovn Client
Currently the event processing goroutine is created after
go-ovn.NewClient() is called. The go-ovn.Client constructor populates
the initial cache content, triggering callbacks that need processing.

If more than 50 callbacks are triggered (the size of the buffered
channel), the have a deadlock.

Fix that by creating the worker goroutine first

Signed-off-by: Adrian Moreno <amorenoz@redhat.com>